requirements.txt
matplotlib==3.8.4
numpy==2.3.1
onnxruntime==1.19.2
onnxruntime_gpu==1.19.2
opencv_python==4.9.0.80
opencv_python_headless==4.10.0.84
PyYAML==6.0.2
PyYAML==6.0.2
scipy==1.16.0
skimage==0.0
torch==2.6.0
tqdm==4.65.2
